graphics:   - implement light caster range in d3d11 default shaders   - fix issue where morph animation wasn't updated because the entity was culled by the shadowcaster frustum   - increase default render window size general:   - change code analysis rule set in visual studio to NativeMinimumRules   - fix potential issues with uninitialised members

+++ code/graphics/subentity.cpp
@@ -60,7 +60,7 @@ SubEntity::SetMaterial(Material* const mat)
     {
         if (VAT_MORPH == this->GetSubMesh()->GetVertexAnimationType())
         {
-            unsigned int i;
+            unsigned short i;
             for (i = 0U; i < this->material->GetNumPasses(); ++i)
             {
                 this->material->GetPass(i)->SetMorphAnimationIncluded(true);
